chore: #31 데일리 투두 인증 생성 API 문서 수정
