+++ src/docs/asciidoc/dailytodo/daily-todo.adoc
@@ -2,7 +2,7 @@
 == 데일리 투두 API
 
 [[create-daily-todos]]
-=== "데일리 투두 생성 API"
+=== 데일리 투두 생성
 
 ==== 개발 상태
 |===
@@ -24,31 +24,8 @@ include::{snippets}/daily-todo-controller-docs-test/create-daily-todos/request-f
 include::{snippets}/daily-todo-controller-docs-test/create-daily-todos/http-response.adoc[]
 include::{snippets}/daily-todo-controller-docs-test/create-daily-todos/response-fields.adoc[]
 
-[[certify-daily-todo]]
-=== 데일리 투두 수행 인증
-
-==== 개발 상태
-|===
-| 환경 | 구현 여부
-
-| 개발
-| O
-
-| 운영
-| X
-|===
-
-==== HTTP Request
-include::{snippets}/daily-todo-controller-docs-test/certify-daily-todo/http-request.adoc[]
-include::{snippets}/daily-todo-controller-docs-test/certify-daily-todo/path-parameters.adoc[]
-include::{snippets}/daily-todo-controller-docs-test/certify-daily-todo/request-fields.adoc[]
-
-==== HTTP Response
-include::{snippets}/daily-todo-controller-docs-test/certify-daily-todo/http-response.adoc[]
-include::{snippets}/daily-todo-controller-docs-test/certify-daily-todo/response-fields.adoc[]
-
 [[get-yesterday-daily-todos]]
-=== 참여중인 특정 챌린지 그룹에서 어제 본인이 작성한 투두 내용 전체 조회 API
+=== 참여중인 특정 챌린지 그룹에서 어제 본인이 작성한 투두 내용 전체 조회
 
 ==== 개발 상태
 |===
@@ -70,7 +47,7 @@ include::{snippets}/daily-todo-controller-docs-test/get-yesterday-daily-todos/ht
 include::{snippets}/daily-todo-controller-docs-test/get-yesterday-daily-todos/response-fields.adoc[]
 
 [[get-my-daily-todos]]
-=== 참여중인 특정 챌린지 그룹에서 내 데일리 투두 전체 조회 API (투두 작성 날짜만 입력)
+=== 참여중인 특정 챌린지 그룹에서 내 데일리 투두 전체 조회 (투두 작성 날짜만 입력)
 
 ==== 개발 상태
 |===
@@ -93,7 +70,7 @@ include::{snippets}/daily-todo-controller-docs-test/get-my-daily-todos-with-cert
 include::{snippets}/daily-todo-controller-docs-test/get-my-daily-todos-with-certification-input-date/response-fields.adoc[]
 
 [[get-my-daily-todos-with-daily-todo-status]]
-=== 참여중인 특정 챌린지 그룹에서 내 데일리 투두 전체 조회 API (투두 작성 날짜 & 투두 상태 입력)
+=== 참여중인 특정 챌린지 그룹에서 내 데일리 투두 전체 조회 (투두 작성 날짜 & 투두 상태 입력)
 
 ==== 개발 상태
 |===
