+++ src/main/java/site/dogether/dailytodo/controller/DailyTodoController.java
@@ -5,7 +5,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import site.dogether.auth.resolver.Authenticated;
 import site.dogether.common.controller.response.ApiResponse;
-import site.dogether.dailytodo.controller.request.CertifyDailyTodoRequest;
 import site.dogether.dailytodo.controller.request.CreateDailyTodosRequest;
 import site.dogether.dailytodo.controller.response.GetMyDailyTodosResponse;
 import site.dogether.dailytodo.controller.response.GetYesterdayDailyTodosResponse;
@@ -19,13 +18,13 @@ import java.util.List;
 import static site.dogether.dailytodo.controller.response.DailyTodoSuccessCode.*;
 
 @RequiredArgsConstructor
-@RequestMapping("/api")
+@RequestMapping("/api/challenge-groups/{groupId}")
 @RestController
 public class DailyTodoController {
 
     private final DailyTodoService dailyTodoService;
 
-    @PostMapping("/challenge-groups/{groupId}/todos")
+    @PostMapping("/todos")
     public ResponseEntity<ApiResponse<Void>> createDailyTodos(
         @Authenticated Long memberId,
         @PathVariable Long groupId,
@@ -35,17 +34,7 @@ public class DailyTodoController {
         return ResponseEntity.ok(ApiResponse.success(CREATE_DAILY_TODOS));
     }
 
-    @PostMapping("/todos/{todoId}/certify")
-    public ResponseEntity<ApiResponse<Void>> certifyDailyTodo(
-        @Authenticated Long memberId,
-        @PathVariable final Long todoId,
-        @RequestBody final CertifyDailyTodoRequest request
-    ) {
-        dailyTodoService.certifyDailyTodo(memberId, todoId, request.content(), request.mediaUrls());
-        return ResponseEntity.ok(ApiResponse.success(CERTIFY_DAILY_TODO));
-    }
-
-    @GetMapping("/challenge-groups/{groupId}/my-yesterday-todos")
+    @GetMapping("/my-yesterday-todos")
     public ResponseEntity<ApiResponse<GetYesterdayDailyTodosResponse>> getYesterdayDailyTodos(
         @Authenticated Long memberId,
         @PathVariable Long groupId
@@ -55,7 +44,7 @@ public class DailyTodoController {
         return ResponseEntity.ok(ApiResponse.successWithData(GET_YESTERDAY_DAILY_TODOS, response));
     }
 
-    @GetMapping("/challenge-groups/{groupId}/my-todos")
+    @GetMapping("/my-todos")
     public ResponseEntity<ApiResponse<GetMyDailyTodosResponse>> getMyDailyTodosWithCertification(
         @Authenticated Long memberId,
         @PathVariable final Long groupId,
