--- src/main/java/site/dogether/dailytodo/controller/request/CertifyDailyTodoRequest.java
@@ -1,6 +1,0 @@
-package site.dogether.dailytodo.controller.request;
-
-import java.util.List;
-
-public record CertifyDailyTodoRequest(String content, List<String> mediaUrls) {
-}
