+++ src/main/java/site/dogether/dailytodocertification/controller/DailyTodoCertificationController.java
@@ -5,6 +5,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import site.dogether.auth.resolver.Authenticated;
 import site.dogether.common.controller.response.ApiResponse;
+import site.dogether.dailytodocertification.controller.request.CertifyDailyTodoRequest;
+import site.dogether.dailytodo.service.DailyTodoService;
 import site.dogether.dailytodocertification.controller.request.ReviewDailyTodoCertificationRequest;
 import site.dogether.dailytodocertification.controller.response.DailyTodoCertificationResponse;
 import site.dogether.dailytodocertification.controller.response.GetDailyTodoCertificationByIdResponse;
@@ -16,16 +18,28 @@ import java.util.List;
 
 import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toList;
+import static site.dogether.dailytodo.controller.response.DailyTodoSuccessCode.CERTIFY_DAILY_TODO;
 import static site.dogether.dailytodocertification.controller.response.DailyTodoCertificationSuccessCode.*;
 
 @RequiredArgsConstructor
-@RequestMapping("/api/todo-certifications")
+@RequestMapping("/api")
 @RestController
 public class DailyTodoCertificationController {
 
+    private final DailyTodoService dailyTodoService;
     private final DailyTodoCertificationService dailyTodoCertificationService;
 
-    @PostMapping("{todoCertificationId}/review")
+    @PostMapping("/todos/{todoId}/certify")
+    public ResponseEntity<ApiResponse<Void>> certifyDailyTodo(
+        @Authenticated Long memberId,
+        @PathVariable final Long todoId,
+        @RequestBody final CertifyDailyTodoRequest request
+    ) {
+        dailyTodoService.certifyDailyTodo(memberId, todoId, request.content(), request.mediaUrl());
+        return ResponseEntity.ok(ApiResponse.success(CERTIFY_DAILY_TODO));
+    }
+
+    @PostMapping("/todo-certifications/{todoCertificationId}/review")
     public ResponseEntity<ApiResponse<Void>> reviewDailyTodoCertification(
         @Authenticated Long memberId,
         @PathVariable Long todoCertificationId,
@@ -40,7 +54,7 @@ public class DailyTodoCertificationController {
         return ResponseEntity.ok(ApiResponse.success(REVIEW_DAILY_TODO_CERTIFICATION));
     }
 
-    @GetMapping("/pending-review")
+    @GetMapping("/todo-certifications/pending-review")
     public ResponseEntity<ApiResponse<GetDailyTodoCertificationsForReviewResponse>> getDailyTodoCertificationsForReview(
         @Authenticated Long memberId
     ) {
@@ -52,7 +66,7 @@ public class DailyTodoCertificationController {
         return ResponseEntity.ok(ApiResponse.successWithData(GET_DAILY_TODO_CERTIFICATIONS_FOR_REVIEW, response));
     }
 
-    @GetMapping("/{todoCertificationId}")
+    @GetMapping("/todo-certifications/{todoCertificationId}")
     public ResponseEntity<ApiResponse<GetDailyTodoCertificationByIdResponse>> getDailyTodoCertificationById(@PathVariable Long todoCertificationId) {
         final DailyTodoCertificationDto dailyTodoCertification = dailyTodoCertificationService.findTodoCertificationById(todoCertificationId);
         final GetDailyTodoCertificationByIdResponse response = new GetDailyTodoCertificationByIdResponse(DailyTodoCertificationResponse.of(dailyTodoCertification));
