+++ src/main/java/site/dogether/dailytodocertification/controller/request/CertifyDailyTodoRequest.java
@@ -1,0 +1,4 @@
+package site.dogether.dailytodocertification.controller.request;
+
+public record CertifyDailyTodoRequest(String content, String mediaUrl) {
+}
