+++ src/test/java/site/dogether/docs/dailytodo/DailyTodoControllerDocsTest.java
@@ -7,7 +7,6 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import site.dogether.challengegroup.entity.ChallengeGroup;
 import site.dogether.challengegroup.entity.ChallengeGroupStatus;
 import site.dogether.dailytodo.controller.DailyTodoController;
-import site.dogether.dailytodo.controller.request.CertifyDailyTodoRequest;
 import site.dogether.dailytodo.controller.request.CreateDailyTodosRequest;
 import site.dogether.dailytodo.entity.DailyTodo;
 import site.dogether.dailytodo.entity.DailyTodoStatus;
@@ -77,48 +76,7 @@ public class DailyTodoControllerDocsTest extends RestDocsSupport {
                         .description("응답 메시지")
                         .type(JsonFieldType.STRING))));
     }
-    
-    @DisplayName("데일리 투두 수행 인증 API")
-    @Test        
-    void certifyDailyTodo() throws Exception {
-        final long todoId = 1L;
-        final CertifyDailyTodoRequest request = new CertifyDailyTodoRequest(
-            "이 노력, 땀 그 모든것이 내 노력의 증거입니다. 양심 있으면 인정 누르시죠.",
-            List.of(
-                "https://dogether-bucket-dev.s3.ap-northeast-2.amazonaws.com/daily-todo-proof-media/mock/e1.png",
-                "https://dogether-bucket-dev.s3.ap-northeast-2.amazonaws.com/daily-todo-proof-media/mock/e2.png"
-            )
-        );
 
-        mockMvc.perform(
-                post("/api/todos/{todoId}/certify", todoId)
-                    .header("Authorization", "Bearer access_token")
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .content(convertToJson(request)))
-            .andExpect(status().isOk())
-            .andDo(createDocument(
-                pathParameters(
-                    parameterWithName("todoId")
-                        .description("데일리 투두 id")
-                        .attributes(constraints("존재하는 데일리 투두 id만 입력 가능"), pathVariableExample(todoId))),
-                requestFields(
-                    fieldWithPath("content")
-                        .description("데일리 투두 인증 본문")
-                        .type(JsonFieldType.STRING)
-                        .attributes(constraints("2 ~ 50 길이 문자열")),
-                    fieldWithPath("mediaUrls")
-                        .description("데일리 투두 인증 미디어 리소스")
-                        .type(JsonFieldType.ARRAY)
-                        .attributes(constraints("MVP에서는 이미지를 올린 S3 URL만 입력 가능"))),
-                responseFields(
-                    fieldWithPath("code")
-                        .description("응답 코드")
-                        .type(JsonFieldType.STRING),
-                    fieldWithPath("message")
-                        .description("응답 메시지")
-                        .type(JsonFieldType.STRING))));
-    }
-    
     @DisplayName("참여중인 특정 챌린지 그룹에서 어제 본인이 작성한 투두 내용 전체 조회 API")
     @Test
     void getYesterdayDailyTodos() throws Exception {
